+++ structurizr-dsl/src/main/java/com/structurizr/dsl/DecisionsParser.java
@@ -6,15 +6,25 @@ import com.structurizr.importer.documentation.DocumentationImporter;
 
 import java.io.File;
 import java.lang.reflect.Constructor;
+import java.util.HashMap;
+import java.util.Map;
 
-final class AdrsParser extends AbstractParser {
+final class DecisionsParser extends AbstractParser {
 
-    private static final String DEFAULT_DECISION_IMPORTER = "com.structurizr.importer.documentation.AdrToolsDecisionImporter";
+    private static final Map<String,String> DECISION_IMPORTERS = new HashMap<>();
 
-    private static final String GRAMMAR = "!adrs <path> <fqn>";
+    private static final String ADRTOOLS_DECISION_IMPORTER = "adrtools";
+    private static final String LOG4BRAINS_DECISION_IMPORTER = "log4brains";
+
+    static {
+        DECISION_IMPORTERS.put(ADRTOOLS_DECISION_IMPORTER, "com.structurizr.importer.documentation.AdrToolsDecisionImporter");
+        DECISION_IMPORTERS.put(LOG4BRAINS_DECISION_IMPORTER, "com.structurizr.importer.documentation.Log4brainsDecisionImporter");
+    }
+
+    private static final String GRAMMAR = "!decisions <path> <type|fqn>";
 
     private static final int PATH_INDEX = 1;
-    private static final int FQN_INDEX = 2;
+    private static final int TYPE_OR_FQN_INDEX = 2;
 
     void parse(WorkspaceDslContext context, File dslFile, Tokens tokens) {
         parse(context, context.getWorkspace(), dslFile, tokens);
@@ -35,7 +45,7 @@ final class AdrsParser extends AbstractParser {
     private void parse(DslContext context, Documentable documentable, File dslFile, Tokens tokens) {
         // !adrs <path>
 
-        if (tokens.hasMoreThan(FQN_INDEX)) {
+        if (tokens.hasMoreThan(TYPE_OR_FQN_INDEX)) {
             throw new RuntimeException("Too many tokens, expected: " + GRAMMAR);
         }
 
@@ -43,9 +53,10 @@ final class AdrsParser extends AbstractParser {
             throw new RuntimeException("Expected: " + GRAMMAR);
         }
 
-        String fullyQualifiedClassName = DEFAULT_DECISION_IMPORTER;
-        if (tokens.includes(FQN_INDEX)) {
-            fullyQualifiedClassName = tokens.get(FQN_INDEX);
+        String fullyQualifiedClassName = DECISION_IMPORTERS.get(ADRTOOLS_DECISION_IMPORTER);
+        if (tokens.includes(TYPE_OR_FQN_INDEX)) {
+            String typeOrFullyQualifiedName = tokens.get(TYPE_OR_FQN_INDEX);
+            fullyQualifiedClassName = DECISION_IMPORTERS.getOrDefault(typeOrFullyQualifiedName, typeOrFullyQualifiedName);
         }
 
         if (dslFile != null) {
@@ -65,14 +76,14 @@ final class AdrsParser extends AbstractParser {
                 DocumentationImporter decisionImporter = (DocumentationImporter)constructor.newInstance();
                 decisionImporter.importDocumentation(documentable, path);
 
-                if (!tokens.includes(FQN_INDEX)) {
+                if (!tokens.includes(TYPE_OR_FQN_INDEX)) {
                     DefaultImageImporter imageImporter = new DefaultImageImporter();
                     imageImporter.importDocumentation(documentable, path);
                 }
             } catch (ClassNotFoundException cnfe) {
-                throw new RuntimeException("Error importing ADRs from " + path.getAbsolutePath() + ": " + fullyQualifiedClassName + " was not found");
+                throw new RuntimeException("Error importing decisions from " + path.getAbsolutePath() + ": " + fullyQualifiedClassName + " was not found");
             } catch (Exception e) {
-                throw new RuntimeException("Error importing ADRs from " + path.getAbsolutePath() + ": " + e.getMessage());
+                throw new RuntimeException("Error importing decisions from " + path.getAbsolutePath() + ": " + e.getMessage());
             }
         }
     }