+++ structurizr-dsl/src/test/java/com/structurizr/dsl/DecisionsParserTests.java
@@ -5,17 +5,17 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
-class AdrsParserTests extends AbstractTests {
+class DecisionsParserTests extends AbstractTests {
 
-    private AdrsParser parser = new AdrsParser();
+    private final DecisionsParser parser = new DecisionsParser();
 
     @Test
     void test_parse_ThrowsAnException_WhenThereAreTooManyTokens() {
         try {
-            parser.parse(new WorkspaceDslContext(), null, tokens("adrs", "path", "fqn", "extra"));
+            parser.parse(new WorkspaceDslContext(), null, tokens("decisions", "path", "fqn", "extra"));
             fail();
         } catch (Exception e) {
-            assertEquals("Too many tokens, expected: !adrs <path> <fqn>", e.getMessage());
+            assertEquals("Too many tokens, expected: !decisions <path> <type|fqn>", e.getMessage());
         }
     }