+++ changelog.md
@@ -5,5 +5,7 @@
 - Removes the deprecated concepts (e.g. location and enterprise.
 - structurizr-core: Adds `Workspace.trim()` to trim a workspace of unused elements (i.e. those not associated with any views).
 - structurizr-client: Removes `StructurizrClient` (use `WorkspaceApiClient` instead).
+- structurizr-import: Adds support for importing decisions managed by Log4brains.
+- structurizr-dsl: Adds `!decisions` as a synonym for `!adrs`.
 - structurizr-assistant: Initial version.